+++ src/Exceptions/VandarBusinessNotConfiguredException.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Zarbinco\LaravelVandar\Exceptions;
+
+class VandarBusinessNotConfiguredException extends VandarException
+{
+    public function __construct()
+    {
+        parent::__construct('Vandar business is not configured.');
+    }
+}