+++ src/Facades/Vandar.php
@@ -16,6 +16,8 @@ use Zarbinco\LaravelVandar\LaravelVandar;
  * @method static \Zarbinco\LaravelVandar\Http\VandarClient client()
  * @method static \Zarbinco\LaravelVandar\Token\TokenManager tokens()
  * @method static \Zarbinco\LaravelVandar\Resources\RawResource raw()
+ * @method static \Zarbinco\LaravelVandar\Resources\BusinessResource business()
+ * @method static \Zarbinco\LaravelVandar\Resources\CustomerResource customers()
  * @method static \Zarbinco\LaravelVandar\DTO\VandarResponse response(array $payload = [], int $status = 200, array $headers = [])
  *
  * @see LaravelVandar